Fix authentication issues

+++ src/main/java/liqour/ravine/service/RatingService.java
@@ -26,7 +26,7 @@ public class RatingService {
         return ratingRepository.findAll();
     }
 
-    public Rating getRatingById(BigInteger id) {
+    public Rating getRatingById(Integer id) {
         return ratingRepository.findById(id).orElse(null);
     }
 
@@ -36,7 +36,7 @@ public class RatingService {
     }
 
     // Delete
-    public void deleteRating(BigInteger id) {
+    public void deleteRating(Integer id) {
         ratingRepository.deleteById(id);
     }
 }
